refactor: address PR #169 review comments (#28)
- JwtAuthenticationFilter: extract BEARER_PREFIX and ROLE_CLAIM constants to eliminate magic numbers/strings
- SecurityProperties: replace @Data with @Getter/@Setter (use only needed annotations)
- SecurityConfig: consolidate shared config into lib/components, remove duplicate copies from auth and api-gateway; add @RequiredArgsConstructor, rename lambda params (auth→requests, session→sessionManagement), rename local var (publicPaths→allowedPublicPaths), add CSRF-disabled comment
- GatewayApplication → GatewayApp: rename class, add @Slf4j and startup log lines
- AuthApplication → AuthApp: rename class, add @Slf4j and startup log lines
