--- packages/auth/src/main/java/com/marzuk/auth/config/SecurityConfig.java
@@ -1,52 +1,0 @@
-package com.marzuk.auth.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.marzuk.components.security.JwtAuthenticationFilter;
-import com.marzuk.components.security.JwtProperties;
-import com.marzuk.components.security.JwtUtil;
-import com.marzuk.components.security.SecurityProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Configuration
-@EnableWebSecurity
-@EnableConfigurationProperties({SecurityProperties.class, JwtProperties.class})
-public class SecurityConfig {
-
-    private final JwtUtil jwtUtil;
-    private final SecurityProperties securityProperties;
-    private final ObjectMapper objectMapper;
-
-    public SecurityConfig(JwtUtil jwtUtil, SecurityProperties securityProperties, ObjectMapper objectMapper) {
-        this.jwtUtil = jwtUtil;
-        this.securityProperties = securityProperties;
-        this.objectMapper = objectMapper;
-    }
-
-    @Bean
-    public JwtAuthenticationFilter jwtAuthenticationFilter() {
-        return new JwtAuthenticationFilter(jwtUtil, securityProperties, objectMapper);
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        String[] publicPaths = securityProperties.getPublicPaths().toArray(new String[0]);
-
-        return http
-                .csrf(AbstractHttpConfigurer::disable)
-                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers(publicPaths).permitAll()
-                        .anyRequest().authenticated()
-                )
-                .addFilterBefore(jwtAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class)
-                .build();
-    }
-}
